requirements.txt
# LangChain & LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Tools & APIs
arxiv>=2.1.0
tavily-python>=0.5.0
duckduckgo-search>=6.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
aiohttp>=3.9.0

# A2A Agent Server
fastapi>=0.115.0
uvicorn>=0.32.0

# PDF Processing
pypdf>=4.0.0

# CLI
rich>=13.0.0
typer>=0.12.0

# MCP Server
fastmcp>=2.0.0

# A2A Protocol
a2a-sdk>=0.2.0
